Exercise 1.3 The c Book page 19 - not sure this is the best solution.
My powers array was in the convert function intially, then I made it a constant so that when you change the size of the ARSIZE you can add to the powers.

My confusion with this exercise was that you are reading in individual numbers from the most significant to the least significant - I had thought of pulling in math.h and using the pow() function rather than generating my own powers function.

This program currently spouts utter rubbish if the number is greater than 99999.